Fix #651
Title attribute for the label should not have been getting set to the
text content of the option element automatically like it was. This was
initially done to give a text value to the click event ui object. The
correct route for a fix should have been to look at the span for the
text value to pass when the event fires rather than always storing it in
the title. This is now being done.