Keyboard Navigation Improvements
ALT-L will now jump from the filter to the first item in the list,
saving a few keys.

SHIFT-TAB in the list of checkboxes will bring focus to the last item in
the header, usually the close button so keyboard navigation can
continue. This is instead of closing the menu.

Got rid of that 'ui-front' addition. It should never have been
committed.