Fix #234
Only try to focus the first label if labels exist, otherwise focus on
the first link in the header.

SHIFT+TAB in the filter input will close the menu to be consistent with
tab behavior elsewhere and to prevent losing cursor.

If there are no options and the user hits tab on the last header
element, the menu will close.