Improve options for header text/links and add Flip All feature
This breaks out the icons to allow them to be something other than the fixed size jQuery UI icons--they could be HTML entities or Font-Awesome icons instead.

Additionally, this adds support for a "flip all" feature that toggles the state of every check box in the menu. This is useful for quickly setting a lot of check-boxes in a long list.

Lastly, tests have been added and modified to accommodate these changes.